+++ src/test/java/org/folio/rest/api/StorageTestSuite.java
@@ -39,6 +39,7 @@ import org.junit.runners.Suite;
   AsyncMigrationTest.class,
   AuditDeleteTest.class,
   BoundWithStorageTest.class,
+  CustomFieldsApiTest.class,
   DereferencedItemStorageTest.class,
   EffectiveLocationMigrationTest.class,
   HoldingsStorageTest.class,